+++ ignite/cmd/scaffold.go
@@ -23,6 +23,7 @@ const (
 	flagNoSimulation = "no-simulation"
 	flagResponse     = "response"
 	flagDescription  = "desc"
+	flagProtoDir     = "proto-dir"
 
 	msgCommitPrefix = "Your saved project changes have not been committed.\nTo enable reverting to your current state, commit your saved changes."
 	msgCommitPrompt = "Do you want to proceed without committing your saved changes"
@@ -131,6 +132,9 @@ with an "--ibc" flag. Note that the default module is not IBC-enabled.
 		NewScaffoldReact(),
 	)
 
+	// Add flags required for the configMigrationPreRunHandler
+	c.PersistentFlags().AddFlagSet(flagSetProtoDir())
+
 	return c
 }
 
@@ -139,10 +143,12 @@ func migrationPreRunHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	session := cliui.New()
+	var (
+		path     = flagGetPath(cmd)
+		protoDir = flagGetProtoDir(cmd)
+		session  = cliui.New()
+	)
 	defer session.End()
-
-	path := flagGetPath(cmd)
 	path, err := filepath.Abs(path)
 	if err != nil {
 		return err
@@ -166,7 +172,7 @@ func migrationPreRunHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	return bufMigrationPreRunHandler(cmd, session, appPath)
+	return bufMigrationPreRunHandler(cmd, session, appPath, protoDir)
 }
 
 func scaffoldType(
@@ -182,6 +188,7 @@ func scaffoldType(
 		withoutSimulation = flagGetNoSimulation(cmd)
 		signer            = flagGetSigner(cmd)
 		appPath           = flagGetPath(cmd)
+		protoDir          = flagGetProtoDir(cmd)
 	)
 
 	var options []scaffolder.AddTypeOption
@@ -206,7 +213,7 @@ func scaffoldType(
 	session := cliui.New(cliui.StartSpinnerWithText(statusScaffolding))
 	defer session.End()
 
-	sc, err := scaffolder.New(cmd.Context(), appPath)
+	sc, err := scaffolder.New(cmd.Context(), appPath, protoDir)
 	if err != nil {
 		return err
 	}