+++ ignite/services/plugin/scaffold_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 
 	"github.com/ignite/cli/v29/ignite/pkg/gocmd"
 	"github.com/ignite/cli/v29/ignite/services/plugin"
@@ -52,7 +52,7 @@ func TestScaffoldedTests(t *testing.T) {
 	// Act
 	err := gocmd.Test(ctx, path, []string{
 		"-timeout",
-		"5m",
+		"10m",
 		"-run",
 		"^TestBar$",
 	})